moved curly parser to its own file, moved curly parser tests to new file and added failure tests
